refactor(ui): rename copy.ts, add barrel, split auth-menu builder
Audit roadmap section 4.1.5 (lib/ui cohesion); zero behavior change.

Rename: lib/ui/copy.ts -> lib/ui/ui-copy.ts via git mv. Premise
deviation: the audit suggested account-display.ts, but the file is the
UI_COPY string-constants table (all user-facing text for the main menu,
account details, oauth flow, return flow, settings, and fallback
prompts) plus formatCheckFlaggedLabel - not account-display logic - so
ui-copy.ts was chosen to match the primary export. Second deviation:
the audit said ~5 importers; there are 14 import sites, all updated:
lib/cli.ts, lib/codex-manager.ts, lib/ui/auth-menu.ts,
lib/codex-manager/{dashboard-display-panel,experimental-settings-schema,
behavior-settings-panel,statusline-settings-panel,theme-settings-panel}.ts,
lib/codex-manager/settings-hub/{experimental,backend,dashboard,index}.ts,
test/documentation.test.ts, test/dashboard-display-panel.test.ts.

Barrel: new lib/ui/index.ts re-exports the surface other modules import
today: ansi (ANSI, isTTY, parseKey, KeyAction), auth-menu
(showAuthMenu, showAccountDetails + action/info types), confirm,
display-width (displayWidth, truncateToWidth), format (paintUiText,
formatUi* helpers, quotaToneFromLeftPercent, UiTextTone), runtime
(get/set/resetUiRuntimeOptions, UiRuntimeOptions), select (select,
MenuItem, SelectOptions), theme (createUiTheme, shouldDisableColor +
theme types), ui-copy (UI_COPY, formatCheckFlaggedLabel). Existing
importers untouched; lib/index.ts does not re-export ui.

Split: auth-menu.ts (724 lines) -> auth-menu-builder.ts (445 lines)
holding the menu-construction layer moved verbatim - the
AccountStatus/AccountInfo/AuthMenuOptions/AuthMenuAction/AccountAction
types and the pure label/badge/row/quota formatting helpers
(mainMenuTitleWithVersion, statusBadge, accountTitle, accountSearchText,
accountRowColor, currentMarkerLabel, formatRelativeTime, formatDate,
formatAccountHint, authMenuFocusKey, plus private quota-bar/duration
helpers). auth-menu.ts (308 lines) keeps the interactive event loops
(showAuthMenu, showAccountDetails, promptSearchQuery), imports from the
builder, and re-exports the moved types; builder does not import
auth-menu.ts (no cycle).

Verified: tsc --noEmit clean; eslint --max-warnings=0 clean on lib/ui/
and every touched file; 16 ui-related vitest suites pass (373 tests).
