+++ lib/codex-manager/behavior-settings-panel.ts
@@ -1,6 +1,6 @@
 import { stdin as input, stdout as output } from "node:process";
 import type { DashboardDisplaySettings } from "../dashboard-settings.js";
-import type { UI_COPY } from "../ui/copy.js";
+import type { UI_COPY } from "../ui/ui-copy.js";
 import { getUiRuntimeOptions } from "../ui/runtime.js";
 import { type MenuItem, select } from "../ui/select.js";
