+++ lib/codex-manager/settings-hub/backend.ts
@@ -1,6 +1,6 @@
 import { loadPluginConfig } from "../../config.js";
 import type { PluginConfig } from "../../types.js";
-import { UI_COPY } from "../../ui/copy.js";
+import { UI_COPY } from "../../ui/ui-copy.js";
 import { getUiRuntimeOptions } from "../../ui/runtime.js";
 import { select } from "../../ui/select.js";
 import { promptBackendCategorySettingsEntry } from "../backend-category-entry.js";
