+++ lib/codex-manager/theme-settings-panel.ts
@@ -4,7 +4,7 @@ import type {
 	DashboardDisplaySettings,
 	DashboardThemePreset,
 } from "../dashboard-settings.js";
-import type { UI_COPY } from "../ui/copy.js";
+import type { UI_COPY } from "../ui/ui-copy.js";
 import { getUiRuntimeOptions } from "../ui/runtime.js";
 import { type MenuItem, select } from "../ui/select.js";
