+++ lib/ui/index.ts
@@ -1,0 +1,39 @@
+export { ANSI, isTTY, parseKey, type KeyAction } from "./ansi.js";
+export {
+	showAccountDetails,
+	showAuthMenu,
+	type AccountAction,
+	type AccountInfo,
+	type AccountStatus,
+	type AuthMenuAction,
+	type AuthMenuOptions,
+} from "./auth-menu.js";
+export { confirm } from "./confirm.js";
+export { displayWidth, truncateToWidth } from "./display-width.js";
+export {
+	formatUiBadge,
+	formatUiHeader,
+	formatUiItem,
+	formatUiKeyValue,
+	formatUiSection,
+	paintUiText,
+	quotaToneFromLeftPercent,
+	type UiTextTone,
+} from "./format.js";
+export {
+	getUiRuntimeOptions,
+	resetUiRuntimeOptions,
+	setUiRuntimeOptions,
+	type UiRuntimeOptions,
+} from "./runtime.js";
+export { select, type MenuItem, type SelectOptions } from "./select.js";
+export {
+	createUiTheme,
+	shouldDisableColor,
+	type UiAccent,
+	type UiColorProfile,
+	type UiGlyphMode,
+	type UiPalette,
+	type UiTheme,
+} from "./theme.js";
+export { UI_COPY, formatCheckFlaggedLabel } from "./ui-copy.js";
